refactor: image sprite 통일
