+++ src/@type/index.d.ts
@@ -32,11 +32,7 @@ declare module Login {
         url?: string;
     }
 
-    interface ImageProps {
-        width: number;
-        height: number;
-        position: string;
-    }
+
 
     interface InputProps {
         inputName: "email" | "name" | "username" | "password" | "id";
@@ -126,5 +122,14 @@ declare module HomeType {
         } | null;
         isCopiedNotification: boolean;
         homeModal: homeModalProps;
+    }
+}
+
+declare module Common {
+    interface ImageProps {
+        width: number;
+        height: number;
+        position: string;
+        url:string;
     }
 }
