+++ src/components/Direct/Section/InboxSection.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 import styled from "styled-components";
-import ImageSprite from "components/Common/DirectSprite";
+import ImageSprite from "components/Common/ImageSprite";
 import Button from "styles/UI/Button";
-import { useAppDispatch } from "../../../app/store/hooks";
-import { openModal } from "../../../app/store/ducks/direct/DirectSlice";
+import { useAppDispatch } from "app/store/hooks";
+import { openModal } from "app/store/ducks/direct/DirectSlice";
+import sprite from "assets/Images/sprite.png";
 
 
 const InboxSectionContainer = styled.div`
@@ -12,9 +13,11 @@ const InboxSectionContainer = styled.div`
   justify-content: center;
   align-items: center;
   height: 100%;
-  .paper-image{
+
+  .paper-image {
     margin-bottom: 10px;
   }
+
   .content-section {
     display: flex;
     flex-direction: column;
@@ -39,31 +42,31 @@ const InboxSectionContainer = styled.div`
   }
 `;
 
-const paperImage: Login.ImageProps = {
+const paperImage: Common.ImageProps = {
     width: 96,
     height: 96,
     position: `-196px -205px`,
+    url: sprite,
 };
 
 
 const InboxSection = () => {
-    const dispatch = useAppDispatch()
+    const dispatch = useAppDispatch();
     return (
         <InboxSectionContainer>
 
             <div className="paper-image">
-                <ImageSprite width={paperImage.width}
-                             height={paperImage.height}
-                             position={paperImage.position} />
+                <ImageSprite {...paperImage}
+                />
             </div>
             <div className="content-section">
                 <span className={"content-title"}>내 메시지</span>
                 <span className={"content-desc"}>친구나 그룹에 비공개 사진과 메시지를 보내보세요.</span>
             </div>
             <div className="button-section">
-                <Button onClick={()=>{
-                    dispatch(openModal("newChat"))
-                }} bgColor="#0095f6" color="#fff" >메시지 보내기</Button>
+                <Button onClick={() => {
+                    dispatch(openModal("newChat"));
+                }} bgColor="#0095f6" color="#fff">메시지 보내기</Button>
             </div>
