+++ src/components/Signup/SignUpForm.tsx
@@ -2,7 +2,8 @@ import styled from "styled-components";
 import FacebookLogin from "components/Common/FacebookLogin";
 import Line from "components/Common/Line";
 import InputAndButton from "./InputAndButton";
-import ImageSprite from "components/Common/LoginSprite";
+import ImageSprite from "components/Common/ImageSprite";
+import sprite from "assets/Images/loginPageSprite.png";
 
 const SignUpFormContainer = styled.div`
     display: flex;
@@ -28,19 +29,18 @@ const SignUpFormContainer = styled.div`
     }
 `;
 
-const instagramImage: Login.ImageProps = {
+const instagramImage: Common.ImageProps = {
     width: 175,
     height: 51,
     position: `0 -130px`,
+    url:sprite
 };
 
 export default function SignUpForm() {
     return (
         <SignUpFormContainer>
             <ImageSprite
-                width={instagramImage.width}
-                height={instagramImage.height}
-                position={instagramImage.position}
+                {...instagramImage}
                 className="logo"
             />
             <form className="signUpForm">
